test: consolidate test files and use require.IsType
Merge all separate *_benchmark_test.go and *_optimize_test.go files into
the existing test files for each package (one test file per processor).
Replace `ok := val.(T); require.True(t, ok)` anti-pattern with
`require.IsType(t, T{}, val)` throughout all new/modified test code.
Add assert.NotSame checks in independence tests to verify maps are not
aliased pointers across events.
